training/requirements.txt
implicit==0.7.2
pandas==2.2.2
scipy==1.13.0
numpy==1.26.4
azure-storage-blob==12.19.0